% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicParam-class.R
\name{musicParam}
\alias{musicParam}
\title{Make new object of class musicParam}
\usage{
musicParam(
  bulkExpression,
  referenceExpression,
  cellScaleFactors = NULL,
  sigma = NULL,
  nu = NULL,
  iter.max = NULL,
  eps = NULL,
  returnInfo = FALSE
)
}
\arguments{
\item{bulkExpression}{Bulk mixed signals matrix of samples, which can be 
matched to single-cell samples.}

\item{referenceExpression}{Signature matrix of cell type-specific signals. 
If not provided, can be computed from a
provided ExpressionSet containing single-cell data.}

\item{cellScaleFactors}{Cell size factor transformations of length equal to 
the K cell types to deconvolve.}

\item{sigma}{Additional argument for algorithm.}

\item{nu}{Additional argument for algorithm.}

\item{iter.max}{Additional argument for algorithm.}

\item{eps}{Additional argument for algorithm.}

\item{returnInfo}{Whether to return metadata and original method outputs 
with predicted proportions.}
}
\value{
New object of class \linkS4class{musicParam}.
}
\description{
Main constructor for class \linkS4class{musicParam}.
}
\details{
Takes standard inputs for the MuSiC algorithm
}
\examples{
exampleDataList <- getDeconvolutionExampleData()
newParam <- musicParam(
exampleDataList[["bulkExpression"]],exampleDataList[["referenceExpression"]],
exampleDataList[["cellScaleFactors"]]
)
deconvolution(newParam)

}
